model db validation

+++ model/DeleteGuideModel.php
@@ -4,6 +4,7 @@
 namespace app\model;
 
 
+use app\core\Application;
 use app\core\BaseModel;
 use Exception;
 
@@ -20,9 +21,23 @@ class DeleteGuideModel extends BaseModel {
         if (!is_numeric($this->guideId)) {
             throw new Exception('Error, guide id is not a number');
         }
+
+        if (!$this->existsInDatabase('guide','id',$this->guideId)) {
+            throw new Exception('Error, guide id not found');
+        }
     }
 
     function delete() {
+        $statement = 'SELECT * FROM guide where id = (?)';
+        $query = $this->prepare($statement);
+        $query->execute([$this->guideId]);
+        $guide = $query->fetch();
+
+        $path = Application::$FILES_PATH . $guide['filename'];
+        if (!unlink($path)) {
+            throw new Exception('Error, file could not be removed');
+        }
+
         $statement = 'DELETE FROM guide WHERE id = (?)';
         $query = $this->prepare($statement);
         $query->execute([$this->guideId]);
